fix: update distance_km type to Float and improve CardRideDetails layout

+++ frontend/src/gql/graphql.ts
@@ -178,7 +178,7 @@ export type Ride = {
   departure_at: Scalars['DateTimeISO']['output'];
   departure_city: Scalars['String']['output'];
   departure_location: Point;
-  distance_km?: Maybe<Scalars['Int']['output']>;
+  distance_km?: Maybe<Scalars['Float']['output']>;
   driver: User;
   duration_min?: Maybe<Scalars['Int']['output']>;
   id: Scalars['ID']['output'];
